BK64: Move collision into its own module
